IIFE implementation
Wraped in IIFE the whole App.
Created pokemonRepository in IIFE alonside with the table to hold all the data passed by funciton add, the pokemonRepository have also getAll function to display it's content and returns functions to use them as methods to add data to the repository. This controls which data should interact within the global scope and prevent accidental change.