Api implementation
Removed unnecessary static code creating the list and replaced it with functions loadList and loadDetails to create dynamic one. Also implemented prototype of a loading message function . Added auto overflow style and fixed height to the app window. Also added polyfills for fetch and promises in order for older browsers and those who are not supporthing them.